откат на предыдущий коммит
