+++ index.html
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
 
-  <script type="module" src="./js/main.js"></script>
+  <script src="./js/main.js"></script>
 
   <title>Кекстаграм</title>
 </head>
